only open websocket when starting streaming
This will allow restarting the stream when some
error happens, without the need for refreshing
the page.

Next step is creating a separate API instead of
the handshake inside the ws, so we don't have the
back and forth all the time and can do the handshake
preemptyively.

This also closes #6